Simplify stklm0 app to inference-only; fix np.isnan crash on nullable Int64
- stklm0/app.py: remove Train & Evaluate section entirely; app now only runs
  Milan BCR/CSM model inference on uploaded STKLM0 CSV; cleaned up helper
  functions, constants, and UI controls that belonged to the removed section
- stklm0/scripts/prepare_stklm0.py: add .astype(float64) after df_days
  construction so col.dt.days nullable Int64 is coerced to plain numpy float
  before np.isnan() is called, preventing TypeError on newer pandas versions

+++ stklm0/scripts/prepare_stklm0.py
@@ -93,7 +93,7 @@ def build_psa_long_stklm0(df: pd.DataFrame, t_max: float = T_MAX, n_psa: int = N
         lambda col: col.dt.days if hasattr(col, "dt") else col.map(
             lambda x: x.days if pd.notna(x) else np.nan
         )
-    )
+    ).astype("float64")   # nullable Int64 → plain float64 so np.isnan works
     df_psa = df[psa_cols].copy().apply(pd.to_numeric, errors="coerce")
 
     neg_mask = df_days < 0
